+++ cypress/e2e/crud-movie-event.cy.ts
@@ -20,8 +20,7 @@ describe('CRUD movie', () => {
     name: spok.string,
     year: spok.number,
     rating: spok.number,
-    director: spok.string,
-    oscar: spok.type('boolean')
+    director: spok.string
   }
 
   let token: string
