+++ cypress/e2e/crud-movie.cy.ts
@@ -17,8 +17,7 @@ describe('CRUD movie', () => {
     name: spok.string,
     year: spok.number,
     rating: spok.number,
-    director: spok.string,
-    oscar: spok.type('boolean')
+    director: spok.string
   }
 
   let token: string
