+++ prisma/schema.prisma
@@ -13,10 +13,9 @@ datasource db {
 }
 
 model Movie {
-  id       Int     @id @default(autoincrement())
+  id       Int    @id @default(autoincrement())
   name     String
   year     Int
   rating   Float
   director String
-  oscar    Boolean
 }
