+++ src/api-docs/openapi-generator.ts
@@ -192,7 +192,7 @@ export const openApiDoc = generator.generateDocument({
   openapi: '3.1.0',
   info: {
     title: 'Movies API',
-    version: '10.0.0',
+    version: '9.0.0',
     description: 'API for managing movies'
   },
   servers: [
