+++ src/movie-adapter.test.ts
@@ -262,15 +262,13 @@ describe('MovieAdapter', () => {
       year: 2020,
       id,
       rating: 7.5,
-      director: 'Christopher Nolan',
-      oscar: true
+      director: 'Christopher Nolan'
     }
     const updateMovieData = {
       name: 'The Dark Knight',
       year: 2008,
       rating: 8.5,
-      director: 'Steven Spielberg',
-      oscar: true
+      director: 'Steven Spielberg'
     }
     const updatedMovie = { id, ...updateMovieData }
