+++ src/test-helpers/factories.ts
@@ -6,8 +6,7 @@ export const generateMovieWithoutId = (): Omit<Movie, 'id'> => {
     name: faker.lorem.words(3), // random 3-word title
     year: faker.date.past({ years: 50 }).getFullYear(), // random year between the past 50 years
     rating: faker.number.float({ min: 1, max: 10, fractionDigits: 1 }), // random rating between 1 and 10 with 1 decimal place
-    director: faker.lorem.words(3),
-    oscar: faker.datatype.boolean()
+    director: faker.lorem.words(3)
   }
 }
 
@@ -17,7 +16,6 @@ export const generateMovieWithId = (): Movie => {
     name: faker.lorem.words(3),
     year: faker.date.past({ years: 50 }).getFullYear(),
     rating: faker.number.float({ min: 1, max: 10, fractionDigits: 1 }),
-    director: faker.lorem.words(3),
-    oscar: faker.datatype.boolean()
+    director: faker.lorem.words(3)
   }
 }
